refactor: share normalizePrivateKey and normalize signNostrEvent pubkey
