+++ utils/index.ts
@@ -3,4 +3,5 @@ export * from './constants.js';
 export * from './conversion.js';
 export * from './formatting.js';
 export * from './pool.js';
-export * from './ephemeral-relay.js'; 
+export * from './ephemeral-relay.js';
+export * from './keys.js';
